Move VersionHelper to VersionHandler
