--- backend/src/Examples/ExampleApp.Examples.Api/Helpers/VersionHelper.cs
@@ -1,16 +1,0 @@
-using System.Reflection;
-
-namespace ExampleApp.Examples.Api.Helpers;
-
-public static class VersionHelper
-{
-    public static readonly string Version;
-
-    static VersionHelper()
-    {
-        var self = Assembly.GetExecutingAssembly();
-        var version = self.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion ?? "0.0.0";
-        var name = self.GetName().Name;
-        Version = $"{name} {version}";
-    }
-}
